training/requirement-clean.txt
langchain
sentence-transformers
faiss-cpu
pgvector
psycopg2-binary

torch
transformers
scikit-learn
pandas
numpy
tqdm
python-dotenv
pymupdf

google-generativeai==0.5.2
